Final Project Upload
This is the final version of my project being uploaded to github from my PC.

Authenticate.php is the page that controls the login functionality and serves as the first page anyone sees.

Game.php is the file that contains my pong game.

Home.php is the page that the user sees after logging in (authenticating).

Logout.php is a file made by the professor that allows a user to logout manually after logging in.

Register.php controls the registration process of the application, includes the sanitation of both the email, username and makes sure the password fulfills specific criteria.

Styles.css controls the style of the html for my .php files.

Get_settings.php is hardcoded settings that are saved in order to show the functionality of loading prior settings.

Save_score.php is the functionality of the score from my game being saved and inputted into my sql table.